requirements.txt
boto==2.42.0
boto3==1.3.1
docopt==0.6.1
troposphere==1.7.0
PyYAML==3.11
